// InnovatecWebFrontend/src/app/shared/models/real-time-pro-solar.ts
export class RealTimeProSolar {
  skey: string;
  corrienteBateria: string;
  corrienteCorriente: string;
  corrientePanel: string;
  fechaActual: string;
  fechaActual1: string;
  hora: string;
  humedad: string;
  irradiancia: string;
  temperatura: string;
  voltajeBateria: string;
  voltajeCargas: string;
  voltajePanel: string;

}